Adding edge case checks to tournament service
Check if player already in tournament when adding

+++ src/main/java/com/ila/checkmatecentral/exceptions/PlayerAlreadyInTournamentException.java
@@ -1,7 +1,7 @@
 package com.ila.checkmatecentral.exceptions;
 
 public class PlayerAlreadyInTournamentException extends RuntimeException {
-    public PlayerAlreadyInTournamentException(Long tournamentId) {
+    public PlayerAlreadyInTournamentException(Integer tournamentId) {
         super("Player is already registered in tournament with ID: " + tournamentId);
     }
 }
